requirements.txt
absl-py==2.1.0
addict==2.4.0
black==24.4.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
einops==0.7.0
filelock==3.13.3
fonttools==4.51.0
fsspec==2024.3.1
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.62.1
h5py==3.10.0
huggingface-hub==0.22.2
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
kiwisolver==1.4.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mmcv==2.1.0
mmcv-full==1.7.2
mmengine==0.10.3
mypy-extensions==1.0.0
numpy==1.22.4
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==24.0
pandas==2.0.3
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.0
protobuf==5.26.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.17.2
pyparsing==3.1.2
PySnooper==1.2.0
python-dateutil==2.9.0.post0
pytorch-ignite==0.4.13
pytorch-msssim==1.0.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scipy==1.7.3
six==1.16.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
timm==0.6.12
tomli==2.0.1
torch==1.8.1+cu101 
torchinfo==1.8.0
TorchSnooper==0.8
torchvision==0.9.1+cu101
tqdm==4.66.1
typing==3.7.4.3
typing_extensions==4.7.1
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
yapf==0.40.2
zipp==3.18.1
